+++ src/cache/mod.rs
@@ -1,5 +1,4 @@
 pub mod arg;
-pub mod args;
 
 use anyhow::{anyhow, Context, Result};
 use dirs::data_dir;